move username to own object
