+++ comp/ConnectButton.jsx
@@ -13,7 +13,6 @@ module.exports = React.createClass({
 	mixins: [SPM.connect('serverStore', '', ['connection', 'users', 'nick'])],
 	render: function(){
 		var onclick, img, label;
-		var userLabel = null;
 		if (this.state.connection === ConnectionState.DISCONNECTED) {
 			onclick = Server.connect;
 			img = require('img/red.png');
@@ -26,29 +25,12 @@ module.exports = React.createClass({
 			onclick = Server.disconnect;
 			img = require('img/green.png');
 			label = 'Connected';
-
-			var user = this.state.users[this.state.nick];
-			var userImg = "";
-			if (user.elo > 0 && user.level >= 0){ //zkls
-				var level = Math.max(1, Math.min(9, Math.floor(10 - 9 * Math.exp(-user.level/60))));
-				var skill = Math.max(0, Math.min(5, Math.floor((user.elo - 1200) / 200)));
-				userImg = require('img/ranks/' + level + '_' + skill + '.png');
-			}
-			else if (user.timeRank >= 0) { //spring
-				var level = user.timeRank + 1;
-				var skill = 2;
-				userImg = require('img/ranks/' + level + '_' + skill + '.png');
-			}
-			userLabel = <div><img src={userImg} /> {this.state.nick}</div>
 		} else { 
 			onclick = null;
 			img = null;
 			label = "[Unknown state, add support]";
 		}
-		// Should probably move the user label into a separate component.
-		// Because now if you click on it you get disconnected.
 		return <button onClick={onclick}><img src={img} /> {label}
-			<div className="userLabel">{userLabel}</div>
 		</button>;
 	}
 });
