+++ style/App.sass
@@ -89,8 +89,6 @@ $downloadBarTransitionDuration: 200ms
 		top: 0
 	.topRight
 		top: $downloadBarHeight
-		.userLabel
-			display: none
 
 // Downloads and Connect buttons, GameInfo status.
 
@@ -111,9 +109,7 @@ $downloadBarTransitionDuration: 200ms
 			top: 0.3em
 			margin-right: 0.3em
 	.userLabel
-		position: absolute
-		top: 2.8em
-		right: 0.7em
+		position: relative
 
 .appContent .topRight .topRightButtons
 	background: #111
